+++ CHANGELOG.md
@@ -1,8 +1,12 @@
+# v0.0.3 - 2018-02-27
+- Allow initial view to be expanded, collapsed or today
+- Add Search (filter) function
+
 # v0.0.2 - 2018-02-24
 - Fix month names
 - Add day of the week and ordinal indicators
-- update when saving new entry
-- remove command search debug
+- Update when saving new entry
+- Remove command search debug
 
 # v0.0.1 - 2018-02-23
 - Initial version